🚧 Some extra notes and settings (read full commit message for musings)
It seems that the maintainability aspect of this is going to become a
problem. I think what is going to have to happen is that there will be a
separate repo with all of these JSON files and a python3 script that can
then be used to generate the syntax file and then this plugin will just
have the 'latest' whenever I run it. It's not ideal, to be sure, but it
does beat having this issue where I have no idea how I'm going to
generate this stuff *on the fly* for all projects ever. Hell, even
running CMake is going to be a nightmare.

Another reason to consider moving this stuff to another file is to
simply allow the eventual vscode, (possibly emacs) and other editors to
have up to date "not shit" files. Hell, even having a highlight.js file
that isn't hot garbage would be an improvement :)

I'll need to figure this out later, but for now I'll keep working on the
vim generation side of these syntax files. If only there was a way to
*for sure* guarantee that people can get all of these as packages.

What would also be nice would be some way to have an opt-in so that the
pack-add! command works just fine for adding ixm and other "large" cmake
libraries